% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQuery.R
\name{dbQuery}
\alias{dbQuery}
\title{Get results from query}
\usage{
dbQuery(conn, statement, ...)
}
\arguments{
\item{conn}{The connection to the database}

\item{statement}{The SQL query statement}

\item{...}{currently ignored}
}
\description{
This function executes the query and returns query results
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHost(),
                 user='username',
                 password='password',
                 catalog='doordastats_snapshot',
                 schema='doordastats_snapshot')

table = dbQuery(conn, "SELECT * FROM table_name")
}
}
