% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doordahost.R
\docType{methods}
\name{DoordaHost}
\alias{DoordaHost}
\alias{dbConnect,DoordaHostDriver-method}
\title{DoordaHost DBI wrapper}
\usage{
DoordaHost()

\S4method{dbConnect}{DoordaHostDriver}(drv, user, password, catalog = "",
  schema = "", ...)
}
\arguments{
\item{drv}{An object created by \code{DoordaHost()}}

\item{user}{user of account}

\item{password}{password of account}

\item{catalog}{catalog to connect to}

\item{schema}{schema to connect to (optional)}

\item{...}{Other options}
}
\description{
DoordaHost DBI wrapper

Wrapper to connect to Doorda Hosted platform.
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHostClient::DoordaHost(),
                 user='username',
                 password='password',
                 catalog='catalog_name',
                 schema='schema_name')

dbQuery(conn, "SELECT * FROM table_name")
}
}
\seealso{
\href{https://github.com/Doorda/Getting-Started}{Getting Started Guide} for more connections options and information.
}
