% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoordaHostConnection.R, R/dbExistsTable.R,
%   R/dbListFields.R
\docType{class}
\name{DoordaHostConnection-class}
\alias{DoordaHostConnection-class}
\alias{show,DoordaHostConnection-method}
\alias{dbExistsTable,DoordaHostConnection,character-method}
\alias{dbListFields,DoordaHostConnection,character-method}
\title{S4 implementation of \code{DBIConnection} for DoordaHost}
\usage{
\S4method{show}{DoordaHostConnection}(object)

\S4method{dbExistsTable}{DoordaHostConnection,character}(conn, name, ...)

\S4method{dbListFields}{DoordaHostConnection,character}(conn, name, ...)
}
\description{
S4 implementation of \code{DBIConnection} for DoordaHost
}
\keyword{internal}
