% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doordahost.R
\docType{class}
\name{DoordaHostDriver-class}
\alias{DoordaHostDriver-class}
\alias{DoordaHostDriver}
\title{DoordaHost driver class.}
\usage{
\S4method{initialize}{DoordaHostDriver}(.Object, ...)
}
\description{
DoordaHost driver class.

Constructor of DoordaHostDriver
}
\keyword{internal}
