% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoordaHostResult.R, R/dbHasCompleted.R,
%   R/dbClearResult.R, R/dbFetch.R, R/dbGetRowCount.R, R/dbGetStatement.R,
%   R/dbIsValid.R, R/dbListFields.R, R/fetch.R
\docType{class}
\name{DoordaHostResult-class}
\alias{DoordaHostResult-class}
\alias{show,DoordaHostResult-method}
\alias{dbHasCompleted,DoordaHostResult-method}
\alias{dbClearResult,DoordaHostResult-method}
\alias{dbFetch,DoordaHostResult,integer-method}
\alias{dbFetch,DoordaHostResult,numeric-method}
\alias{dbFetch,DoordaHostResult,missing-method}
\alias{dbGetRowCount,DoordaHostResult-method}
\alias{dbGetStatement,DoordaHostResult-method}
\alias{dbIsValid,DoordaHostResult-method}
\alias{dbListFields,DoordaHostResult,missing-method}
\alias{fetch,DoordaHostResult,integer-method}
\alias{fetch,DoordaHostResult,numeric-method}
\alias{fetch,DoordaHostResult,missing-method}
\title{An S4 class to represent a DoordaHost Result}
\usage{
\S4method{show}{DoordaHostResult}(object)

\S4method{dbHasCompleted}{DoordaHostResult}(res, ...)

\S4method{dbClearResult}{DoordaHostResult}(res, ...)

\S4method{dbFetch}{DoordaHostResult,integer}(res, n = -1, ...)

\S4method{dbFetch}{DoordaHostResult,numeric}(res, n = -1, ...)

\S4method{dbFetch}{DoordaHostResult,missing}(res, n = -1, ...)

\S4method{dbGetRowCount}{DoordaHostResult}(res, ...)

\S4method{dbGetStatement}{DoordaHostResult}(res, ...)

\S4method{dbIsValid}{DoordaHostResult}(dbObj, ...)

\S4method{dbListFields}{DoordaHostResult,missing}(conn, name)

\S4method{fetch}{DoordaHostResult,integer}(res, n = -1, ...)

\S4method{fetch}{DoordaHostResult,numeric}(res, n = -1, ...)

\S4method{fetch}{DoordaHostResult,missing}(res)
}
\description{
An S4 class to represent a DoordaHost Result
}
\section{Slots}{

\describe{
\item{\code{statement}}{The SQL statement sent to the database}

\item{\code{connection}}{The connection object associated with the result}

\item{\code{cursor}}{An internal implementation detail for keeping track of
what stage a request is in}
}}

\keyword{internal}
