% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableStats.R
\name{getTableStats}
\alias{getTableStats}
\title{Get Table Stats}
\usage{
getTableStats(conn, table_name, ...)
}
\arguments{
\item{conn}{The connection to the database}

\item{table_name}{The name of the table in the catalogs-schema}

\item{...}{currently ignored}
}
\description{
This function returns the number of rows in the input table
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHost(),
                 user='username',
                 password='password',
                 catalog='doordastats_snapshot',
                 schema='doordastats_snapshot')

catalogs = getTableStats(conn)
}
}
