% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{isConnected}
\alias{isConnected}
\title{Check DoordaHost connection}
\usage{
isConnected(connection)
}
\arguments{
\item{connection}{The connection to the database}
}
\description{
This function sends a ping to DoordaHost to check if connection is successful
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHostClient::DoordaHost(),
                 user='username',
                 password='password',
                 catalog='doordastats_snapshot',
                 schema='doordastats_snapshot')

isConnected(conn)
}
}
