% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{showCatalogs}
\alias{showCatalogs}
\title{List all Catalogs}
\usage{
showCatalogs(connection)
}
\arguments{
\item{connection}{The connection to the database}
}
\description{
This function returns a list of all catalogs
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHostClient::DoordaHost(),
                 user='username',
                 password='password',
                 catalog='catalog_name',
                 schema='schema_name')

catalog_list = showCatalogs(conn)
}
}
