% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSchemas.R
\name{showSchemas}
\alias{showSchemas}
\title{List all Schemas in Catalog}
\usage{
showSchemas(conn, ...)
}
\arguments{
\item{conn}{The connection to the database}

\item{...}{currently ignored}
}
\description{
This function returns a list of all Schemas in defined Catalog
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHost(),
                 user='username',
                 password='password',
                 catalog='doordastats_snapshot',
                 schema='doordastats_snapshot')

catalogs = showSchemas(conn)
}
}
