% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTables.R
\name{showTables}
\alias{showTables}
\title{List Tables in Catalog/Schema}
\usage{
showTables(conn, pattern, ...)
}
\arguments{
\item{conn}{The connection to the database}

\item{pattern}{optional SQL pattern for filtering table names,
e.g. \sQuote{\%test\%}}

\item{...}{currently ignored}
}
\description{
This function returns a list of all tables in defined Catalog/Schema
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHost(),
                 user='username',
                 password='password',
                 catalog='doordastats_snapshot',
                 schema='doordastats_snapshot')

table = showTables(conn, "eng_\%")
}
}
