% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{dbQuery}
\alias{dbQuery}
\title{Get results from query}
\usage{
dbQuery(connection, query)
}
\arguments{
\item{connection}{The connection to the database}

\item{query}{Query statement}
}
\description{
This function executes the query and returns query results
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHostClient::DoordaHost(),
                 user='username',
                 password='password',
                 catalog='catalog_name',
                 schema='schema_name')

catalog_list = dbQuery(conn, "SELECT * FROM table_name")
}
}
