% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoordaHost.R, R/dbConnect.R, R/dbDisconnect.R
\name{DoordaHost}
\alias{DoordaHost}
\alias{dbConnect,DoordaHostDriver-method}
\alias{dbDisconnect,DoordaHostConnection-method}
\title{Connect to a Doorda database}
\usage{
DoordaHost(...)

\S4method{dbConnect}{DoordaHostDriver}(drv, catalog, schema, user, password, ...)

\S4method{dbDisconnect}{DoordaHostConnection}(conn)
}
\arguments{
\item{...}{currently ignored}

\item{drv}{A driver object generated by \code{\link{DoordaHost}}}

\item{catalog}{The catalog to be used}

\item{schema}{The schema to be used}

\item{user}{The current user}

\item{password}{The current users password}

\item{conn}{A \code{\linkS4class{DoordaHostConnection}} object}
}
\value{
[DoordaHost] A \code{\linkS4class{DoordaHostDriver}} object

[dbConnect] A \code{\linkS4class{DoordaHostConnection}} object

[dbDisconnect] A \code{\link{logical}} value indicating success
}
\description{
Connect to a Doorda database
}
\examples{
\dontrun{
  conn <- dbConnect(DoordaHost(), catalog = 'hive', schema = 'default',
                    user = 'onur', password = 'pwd')
  dbListTables(conn, '\%_iris')
  dbDisconnect(conn)
}
}
