% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoordaHostDriver.R, R/dbUnloadDriver.R
\docType{class}
\name{DoordaHostDriver-class}
\alias{DoordaHostDriver-class}
\alias{show,DoordaHostDriver-method}
\alias{dbUnloadDriver,DoordaHostDriver-method}
\title{An S4 class to represent a DoordaHost Driver (and methods)
It is used purely for dispatch and \code{dbUnloadDriver} is unnecessary}
\usage{
\S4method{show}{DoordaHostDriver}(object)

\S4method{dbUnloadDriver}{DoordaHostDriver}(drv, ...)
}
\description{
An S4 class to represent a DoordaHost Driver (and methods)
It is used purely for dispatch and \code{dbUnloadDriver} is unnecessary
}
\keyword{internal}
