% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{getTableStats}
\alias{getTableStats}
\title{Get table stats}
\usage{
getTableStats(connection, catalog = "", schema = "", table = "")
}
\arguments{
\item{connection}{The connection to the database}

\item{catalog}{Name of Catalog}

\item{schema}{Name of Schema (Usually the same as catalog)}

\item{table}{Name of Table}
}
\description{
This function executes the query and returns query results
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHostClient::DoordaHost(),
                 user='username',
                 password='password',
                 catalog='doordastats_snapshot',
                 schema='doordastats_snapshot')

catalog_list = getTableStats(conn, 
                             catalog="doordastats_snapshot", 
                             schema="doordastats_snapshot", 
                             table="eng_area_household_car_van")
}
}
