% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showCatalogs.R
\name{showCatalogs}
\alias{showCatalogs}
\title{List all Catalogs}
\usage{
showCatalogs(conn, ...)
}
\arguments{
\item{conn}{The connection to the database}

\item{...}{currently ignored}
}
\description{
This function returns a list of all Catalogs in defined Catalog/Schema
}
\examples{
\dontrun{
require(DBI)
conn <- dbConnect(DoordaHost(),
                 user='username',
                 password='password',
                 catalog='doordastats_snapshot',
                 schema='doordastats_snapshot')

catalogs = showCatalogs(conn)
}
}
